[consensus] move decoupled execution out from experimental
