+++ consensus/consensus-types/src/lib.rs
@@ -11,7 +11,7 @@ pub mod common;
 pub mod delayed_qc_msg;
 pub mod epoch_retrieval;
 pub mod executed_block;
-pub mod experimental;
+pub mod pipeline;
 pub mod proof_of_store;
 pub mod proposal_ext;
 pub mod proposal_msg;
