+++ consensus/src/dag/commit_signer.rs
@@ -1,6 +1,6 @@
 // Copyright © Aptos Foundation
 
-use crate::experimental::signing_phase::CommitSignerProvider;
+use crate::pipeline::signing_phase::CommitSignerProvider;
 use aptos_crypto::bls12381;
 use aptos_types::validator_signer::ValidatorSigner;
 use std::sync::Arc;
