+++ consensus/src/pipeline/pipeline_phase.rs
@@ -4,7 +4,7 @@
 
 use crate::{
     counters::BUFFER_MANAGER_PHASE_PROCESS_SECONDS,
-    experimental::buffer_manager::{Receiver, Sender},
+    pipeline::buffer_manager::{Receiver, Sender},
 };
 use aptos_logger::debug;
 use async_trait::async_trait;
